+++ src/main/java/proguard/classfile/util/ClassUtil.java
@@ -112,6 +112,7 @@ public class ClassUtil
             externalClassVersion.equals(JavaVersionConstants.CLASS_VERSION_15)  ? VersionConstants.CLASS_VERSION_15  :
             externalClassVersion.equals(JavaVersionConstants.CLASS_VERSION_16)  ? VersionConstants.CLASS_VERSION_16  :
             externalClassVersion.equals(JavaVersionConstants.CLASS_VERSION_17)  ? VersionConstants.CLASS_VERSION_17  :
+            externalClassVersion.equals(JavaVersionConstants.CLASS_VERSION_18)  ? VersionConstants.CLASS_VERSION_18  :
                                                                                   0;
     }
 
@@ -142,6 +143,7 @@ public class ClassUtil
             case VersionConstants.CLASS_VERSION_15:  return JavaVersionConstants.CLASS_VERSION_15;
             case VersionConstants.CLASS_VERSION_16:  return JavaVersionConstants.CLASS_VERSION_16;
             case VersionConstants.CLASS_VERSION_17:  return JavaVersionConstants.CLASS_VERSION_17;
+            case VersionConstants.CLASS_VERSION_18:  return JavaVersionConstants.CLASS_VERSION_18;
             default:                                 return null;
         }
     }