+++ src/main/java/proguard/classfile/util/kotlin/KotlinMetadataInitializer.java
@@ -22,6 +22,7 @@ import kotlinx.metadata.internal.metadata.jvm.deserialization.JvmMetadataVersion
 import kotlinx.metadata.jvm.JvmFieldSignature;
 import kotlinx.metadata.jvm.JvmMethodSignature;
 import kotlinx.metadata.jvm.*;
+import org.jetbrains.annotations.Nullable;
 import proguard.classfile.*;
 import proguard.classfile.attribute.annotation.*;
 import proguard.classfile.attribute.annotation.visitor.*;
@@ -612,6 +613,12 @@ implements ClassVisitor,
             }
 
             @Override
+            public void visitJvmFlags(int flags)
+            {
+                setClassJvmFlags(kotlinClassKindMetadata.flags, flags);
+            }
+
+            @Override
             public void visitEnd() {}
         }
 
@@ -891,6 +898,11 @@ implements ClassVisitor,
             public void visitSyntheticMethodForAnnotations(JvmMethodSignature jvmMethodSignature)
             {
                 kotlinPropertyMetadata.syntheticMethodForAnnotations = fromKotlinJvmMethodSignature(jvmMethodSignature);
+            }
+
+            @Override
+            public void visitSyntheticMethodForDelegate(JvmMethodSignature jvmMethodSignature) {
+                kotlinPropertyMetadata.syntheticMethodForDelegate = fromKotlinJvmMethodSignature(jvmMethodSignature);
             }
 
             @Override
@@ -1898,6 +1910,13 @@ implements ClassVisitor,
     {
         flags.isMovedFromInterfaceCompanion =
             JvmFlag.Property.IS_MOVED_FROM_INTERFACE_COMPANION.invoke(jvmFlags);
+    }
+
+
+    private void setClassJvmFlags(KotlinClassFlags flags, int jvmFlags)
+    {
+        flags.hasMethodBodiesInInterface    = JvmFlag.Class.HAS_METHOD_BODIES_IN_INTERFACE.invoke(jvmFlags);
+        flags.isCompiledInCompatibilityMode = JvmFlag.Class.IS_COMPILED_IN_COMPATIBILITY_MODE.invoke(jvmFlags);
     }